Remove duplicate component
