--- src/components/voucher/numeric_input.tsx
@@ -1,181 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import { numberWithCommas } from '@/lib/utils/common';
-
-interface INumericInputProps extends React.InputHTMLAttributes<HTMLInputElement> {
-  value: number | undefined;
-  setValue?: React.Dispatch<React.SetStateAction<number | undefined>>;
-  isDecimal?: boolean;
-  hasComma?: boolean; // Info: (20240722 - Liz) 新增逗號顯示
-  triggerWhenChanged?: (value: number, e: React.ChangeEvent<HTMLInputElement>) => void;
-}
-
-const formatDisplayValue = (
-  value: string | number | undefined,
-  isDecimal: boolean | undefined,
-  hasComma: boolean | undefined
-) => {
-  if (value === undefined) return '';
-  let stringValue = value.toString();
-
-  if (!isDecimal) {
-    const intValue = parseInt(stringValue, 10);
-    stringValue = Number.isNaN(intValue) ? '0' : intValue.toString();
-  }
-
-  return hasComma ? numberWithCommas(stringValue) : stringValue;
-};
-
-const NumericInput = ({
-  value,
-  setValue,
-  isDecimal,
-  hasComma,
-  triggerWhenChanged,
-  ...props
-}: INumericInputProps) => {
-  // Info: (20240723 - Liz) displayValue 是顯示在 input 上的顯示值
-  const [displayValue, setDisplayValue] = useState<string>(value ? value.toString() : '');
-
-  // Info: (20240723 - Liz) dbValue 是存入 DB 的儲存值
-  const [dbValue, setDbValue] = useState<number | undefined>(value);
-
-  useEffect(() => {
-    setDisplayValue(formatDisplayValue(value, isDecimal, hasComma));
-  }, [value, hasComma, isDecimal]);
-
-  useEffect(() => {
-    if (setValue) setValue(dbValue);
-  }, [dbValue, setValue]);
-
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const inputValue = event.target.value;
-
-    // Info: (20240723 - Liz) 整理輸入的值
-    const sanitizedValue =
-      inputValue
-        .replace(/^0+/, '') // Info: (20240723 - Liz) 移除開頭的零
-        .replace(/[^0-9.]/g, '') // Info: (20240723 - Liz) 移除非數字和小數點字符
-        .replace(/(\..*)\./g, '$1') || '0'; // Info: (20240723 - Liz) 只允許一個小數點
-
-    // Info: (20240723 - Liz) 轉換成數值 (整數或浮點數) 為了存入 DB
-    const numericValue = isDecimal
-      ? parseFloat(sanitizedValue) // Info: (20240723 - Liz) 如果解析失敗，會是 NaN
-      : parseInt(sanitizedValue, 10); // Info: (20240723 - Liz) 如果解析失敗，會是 NaN
-
-    // Info: (20240723 - Liz) 處理 NaN 的情況
-    const validNumericValue = Number.isNaN(numericValue) ? 0 : numericValue;
-
-    // Info: (20240723 - Liz) 根據 isDecimal 和 hasComma 的值來決定顯示值的格式
-    const formattedDisplayValue = formatDisplayValue(sanitizedValue, isDecimal, hasComma);
-
-    // Info: (20240723 - Liz) 更新儲存值
-    setDbValue(validNumericValue);
-
-    // Info: (20240723 - Liz) 更新顯示值
-    setDisplayValue(formattedDisplayValue);
-
-    if (triggerWhenChanged) {
-      triggerWhenChanged(validNumericValue, event);
-    }
-  };
-
-  // Info: (20250306 - Julian) 處理在中文輸入法下，填入數字的情況
-  function convertInput(event: React.KeyboardEvent<HTMLInputElement>) {
-    // Info: (20250311 - Julian) 如果按下 tab: 執行預設行為
-    if (event.code === 'Tab') {
-      return;
-    }
-
-    event.preventDefault(); // Info: (20250306 - Julian) 阻止預設事件
-
-    let temp = displayValue; // Info: (20250306 - Julian) 取得目前顯示值
-    let code = ''; // Info: (20250306 - Julian) 按鍵 code
-
-    const input = event.currentTarget; // Info: (20250306 - Julian) 取得 input 元件
-    const cursorPos = input.selectionStart ?? displayValue.length; // Info: (20250306 - Julian) 取得當前游標位置
-
-    if (event.code.indexOf('Digit') > -1) {
-      // Info: (20250306 - Julian) 如果按下的是數字鍵
-      code = event.code.slice(-1); // Info: (20250306 - Julian) 取得數字鍵的值
-      temp = temp.slice(0, cursorPos) + code + temp.slice(cursorPos); // Info: (20250306 - Julian) 插入數字
-
-      // Info: (20250306 - Julian) 變更顯示值
-      handleChange({ target: { value: temp } } as React.ChangeEvent<HTMLInputElement>);
-    } else if (event.code === 'Backspace' && cursorPos > 0) {
-      // Info: (20250306 - Julian) 如果按下 backspace
-      // Info: (20250306 - Julian) 刪除游標前一個字
-      temp = temp.slice(0, cursorPos - 1) + temp.slice(cursorPos);
-      // Info: (20250306 - Julian) 變更顯示值
-      handleChange({ target: { value: temp } } as React.ChangeEvent<HTMLInputElement>);
-
-      // Info: (20250306 - Julian) 變更游標位置
-      requestAnimationFrame(() => {
-        // Info: (20250306 - Julian) 在 displayValue 歸零前，游標向左移動
-        if (displayValue === '0') {
-          input.setSelectionRange(0, 0);
-        } else input.setSelectionRange(cursorPos - 1, cursorPos - 1);
-      });
-    } else if (event.code === 'Delete' && cursorPos < temp.length) {
-      // Info: (20250306 - Julian) 如果按下 delete
-      // Info: (20250306 - Julian) 刪除游標後一個字
-      temp = temp.slice(0, cursorPos) + temp.slice(cursorPos + 1);
-      // Info: (20250306 - Julian) 變更顯示值
-      handleChange({ target: { value: temp } } as React.ChangeEvent<HTMLInputElement>);
-
-      // Info: (20250306 - Julian) 變更游標位置
-      requestAnimationFrame(() => {
-        input.setSelectionRange(cursorPos, cursorPos); // Info: (20250306 - Julian) 游標不動
-      });
-    } else if (event.code === 'ArrowLeft') {
-      // Info: (20250306 - Julian) 如果按下左方向鍵
-      requestAnimationFrame(() => {
-        // Info: (20250306 - Julian) 游標左移
-        input.setSelectionRange(Math.max(cursorPos - 1, 0), Math.max(cursorPos - 1, 0));
-      });
-    } else if (event.code === 'ArrowRight') {
-      // Info: (20250306 - Julian) 如果按下右方向鍵
-      requestAnimationFrame(() => {
-        // Info: (20250306 - Julian) 游標右移
-        input.setSelectionRange(
-          Math.min(cursorPos + 1, temp.length),
-          Math.min(cursorPos + 1, temp.length)
-        );
-      });
-    }
-  }
-
-  // Info: (20240723 - Liz) 處理 displayValue 為空或僅為點的情況
-  const handleBlur = () => {
-    if (!displayValue || displayValue === '.') {
-      setDisplayValue('0');
-      setDbValue(0);
-    }
-  };
-
-  // Info: (20240710 - Julian) 當 input focus 時，如果值為 0，則清空
-  const handleFocus = () => {
-    if (displayValue === '0') {
-      setDisplayValue('');
-    }
-  };
-
-  // Info: (20240710 - Julian) 禁止滾輪改變數值
-  const handleWheel = (e: React.WheelEvent<HTMLInputElement>) => {
-    e.currentTarget.blur();
-  };
-
-  return (
-    <input
-      type="text" // Info: (20240722 - Liz) 保持輸入的 type 為 text 以允許顯示逗號
-      value={displayValue}
-      onChange={handleChange}
-      onFocus={handleFocus}
-      onBlur={handleBlur}
-      onWheel={handleWheel}
-      onKeyDown={convertInput}
-      {...props}
-    />
-  );
-};
-
-export default NumericInput;
